Done: update credit card

+++ model/orderModel.py
@@ -3,13 +3,13 @@ from inf5190.model import models
 from inf5190.model.productModel import Product
 from inf5190.model.shippingInfoModel import ShippingInformation
 from inf5190.model.transactionModel import Transaction
-from peewee import CharField, ForeignKeyField, FloatField, IntegerField, BooleanField
+from peewee import CharField, ForeignKeyField, IntegerField, BooleanField
 
 
 class Order(models.BaseModel):
     product = ForeignKeyField(Product, backref="orders")
     quantity = IntegerField()
-    total_price = FloatField()
+    total_price = IntegerField()
     email = CharField(null=True)
     credit_card = ForeignKeyField(CreditCard, backref="orders", null=True)
     shipping_information = ForeignKeyField(ShippingInformation, backref="orders", null=True)
